feat(config-api): pagination enhacement
