+++ jans-config-api/docs/jans-config-api-swagger-auto.yaml
@@ -3317,17 +3317,17 @@ components:
           $ref: '#/components/schemas/AttributeValidation'
         tooltip:
           type: string
-        adminCanView:
+        adminCanEdit:
           type: boolean
-        userCanAccess:
+        adminCanView:
           type: boolean
         userCanView:
           type: boolean
+        userCanEdit:
+          type: boolean
         adminCanAccess:
           type: boolean
-        adminCanEdit:
-          type: boolean
-        userCanEdit:
+        userCanAccess:
           type: boolean
         whitePagesCanView:
           type: boolean
@@ -3656,8 +3656,6 @@ components:
           format: int32
         displayName:
           type: string
-        tokenBindingSupported:
-          type: boolean
         authenticationMethod:
           type: string
           enum:
@@ -3669,6 +3667,8 @@ components:
           - tls_client_auth
           - self_signed_tls_client_auth
           - none
+        tokenBindingSupported:
+          type: boolean
         baseDn:
           type: string
         inum:
@@ -4773,13 +4773,13 @@ components:
           type: boolean
         internal:
           type: boolean
+        locationPath:
+          type: string
         locationType:
           type: string
           enum:
           - ldap
           - file
-        locationPath:
-          type: string
         baseDn:
           type: string
     ScriptError:
