+++ jans-config-api/server/src/main/java/io/jans/configapi/rest/resource/auth/AttributesResource.java
@@ -9,6 +9,7 @@ package io.jans.configapi.rest.resource.auth;
 import com.github.fge.jsonpatch.JsonPatchException;
 
 import io.jans.configapi.core.model.PatchRequest;
+import io.jans.configapi.core.model.SearchRequest;
 import io.jans.configapi.core.rest.ProtectedApi;
 import io.jans.configapi.service.auth.AttributeService;
 import io.jans.configapi.util.ApiAccessConstants;
@@ -16,7 +17,6 @@ import io.jans.configapi.util.ApiConstants;
 import io.jans.configapi.util.AttributeNames;
 import io.jans.configapi.core.util.Jackson;
 import io.jans.model.GluuAttribute;
-import io.jans.model.SearchRequest;
 import io.jans.orm.model.PagedResult;
 
 import io.swagger.v3.oas.annotations.Operation;
