+++ jans-config-api/server/src/main/java/io/jans/configapi/service/auth/AttributeService.java
@@ -4,7 +4,7 @@ package io.jans.configapi.service.auth;
 import io.jans.model.GluuAttribute;
 import io.jans.as.common.util.AttributeConstants;
 import io.jans.configapi.util.ApiConstants;
-import io.jans.model.SearchRequest;
+import io.jans.configapi.core.model.SearchRequest;
 import io.jans.orm.model.PagedResult;
 import io.jans.orm.model.SortOrder;
 import io.jans.orm.search.filter.Filter;
