+++ jans-config-api/server/src/test/resources/feature/attribute/attributes.feature
@@ -22,6 +22,21 @@ Scenario: Fetch all attributes
 	#And assert response.length != null 
 	#And assert response.length >= 10 
 
+@ignore
+Scenario: Fetch based on filter
+	Given url mainUrl 
+	And print 'accessToken = '+accessToken
+	And print 'issuer = '+issuer
+	And header Authorization = 'Bearer ' + accessToken
+	#And header issuer = issuer 
+	And param limit = 3 	
+	And param pattern = 'edu' 
+	And param startIndex = 1 
+	When method GET 
+	Then status 200 
+	And print response
+	#And assert response.length != null 
+	#And assert response.length >= 10 
 
 Scenario: Fetch the first three attributes 
 	Given url mainUrl
